pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openshield-cli"
version = "0.1.0"
description = "AI-Powered Penetration Testing Automation CLI Tool"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Guardian Team"}
]
keywords = ["penetration-testing", "security", "ai", "automation", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=0.0.5",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "httpx>=0.26.0",
    "asyncio>=3.4.3",
    "aiofiles>=23.2.1",
    "jinja2>=3.1.3",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "textual>=0.47.0",
]


[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
]

[project.scripts]
openshield = "cli.main:app"

[tool.setuptools]
packages = ["cli", "core", "ai", "tools", "reports", "utils", "workflows"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
